fix compiler warnings
git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@33341 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ viterbi/atom.cc
@@ -115,7 +115,7 @@ std::ostream& do_prt(std::ostream& out, const Atom* a, int ilvl)
 
 		ilvl++;
 		size_t lsz = l->get_arity();
-		for (int i=0; i < lsz; i++)
+		for (size_t i=0; i < lsz; i++)
 		{
 			do_prt(out, l->get_outgoing_atom(i), ilvl);
 		}
